+++ fishyrl/losses.py
@@ -21,7 +21,7 @@ def mse_loss(pred: torch.Tensor, target: torch.Tensor, dims: int = 1) -> torch.T
     return (pred - target).square().mean(dim=[-i for i in range(1, dims + 1)])
 
 
-def hungarian_loss(pred: torch.Tensor, target: torch.Tensor, dims: int = 2) -> torch.Tensor:
+def hungarian_loss(pred: torch.Tensor, target: torch.Tensor, dims: int = 2, clip: float = 1e-8) -> torch.Tensor:
     """Compute the Hungarian loss between the predicted and target distributions on the final dimension.
 
     This loss function is not vectorized and is computationally expensive.
@@ -32,6 +32,8 @@ def hungarian_loss(pred: torch.Tensor, target: torch.Tensor, dims: int = 2) -> t
     :type target: torch.Tensor
     :param dims: The number of final dimensions to compute the loss over. (Default: ``1``)
     :type dims: int
+    :param clip: The minimal value of the MSE loss, to prevent NaNs. (Default: ``1e-8``)
+    :type clip: float
     :return: The Hungarian loss.
     :rtype: torch.Tensor
 
@@ -45,7 +47,7 @@ def hungarian_loss(pred: torch.Tensor, target: torch.Tensor, dims: int = 2) -> t
     # Compute MSE loss based on optimal assignment
     ordered_pred = pred[row_ind]
     ordered_target = target[col_ind]
-    return mse_loss(ordered_pred, ordered_target, dims=dims)
+    return mse_loss(ordered_pred, ordered_target, dims=dims).clip(min=clip)
 
 
 def attention_reconstruction_loss(
@@ -53,6 +55,7 @@ def attention_reconstruction_loss(
     entities_target: torch.Tensor,
     existence_logits: torch.Tensor,
     existence_target: torch.Tensor = None,
+    reconstruction_loss_type: str = 'positional',
 ) -> tuple[torch.Tensor, torch.Tensor]:
     """Compute the attention reconstruction loss between the predicted and target distributions.
 
@@ -71,31 +74,43 @@ def attention_reconstruction_loss(
     """
     # Compute existence target if not provided
     if existence_target is None:
-        existence_target = (entities_target.detach().abs().sum(dim=-1) > 0).float()
+        existence_target = (entities_target.detach().isnan().sum(dim=-1) == 0).float()
+
+    # Zero pad entity targets to match entity predictions if needed
+    if entities_target.shape[-2] < entities_pred.shape[-2]:
+        pad_size = entities_pred.shape[-2] - entities_target.shape[-2]
+        entities_target = F.pad(entities_target, (0, 0, 0, pad_size), value=0)  # Pad bottom of second to last dim
 
     # Zero pad existence target to match existence logits if needed
     if existence_target.shape[-1] < existence_logits.shape[-1]:
         pad_size = existence_logits.shape[-1] - existence_target.shape[-1]
         existence_target = F.pad(existence_target, (0, pad_size), value=0)  # Pad bottom of last dim
 
-    # Get batch dims and flatten predictions
-    batch_dims = entities_pred.shape[:-2]
-    entities_pred_flat = entities_pred.view(-1, *entities_pred.shape[-2:])
-    entities_target_flat = entities_target.view(-1, *entities_target.shape[-2:])
-    existence_target_flat = existence_target.view(-1, existence_target.shape[-1])
+    # Hungarian reconstruction loss for non-positional decoder outputs
+    if reconstruction_loss_type == 'hungarian':
+        # Get batch dims and flatten predictions
+        batch_dims = entities_pred.shape[:-2]
+        entities_pred_flat = entities_pred.view(-1, *entities_pred.shape[-2:])
+        entities_target_flat = entities_target.view(-1, *entities_target.shape[-2:])
+        existence_target_flat = existence_target.view(-1, existence_target.shape[-1])
 
-    # Compute entity reconstruction loss
-    reconstruction_loss = torch.zeros(entities_pred_flat.shape[0], device=entities_pred.device)
-    for i in range(entities_pred_flat.shape[0]):
-        # Extract pred and target based on actual number of entities
-        num_entities = existence_target_flat[i].sum().int().item()
-        pred = entities_pred_flat[i][:num_entities]
-        target = entities_target_flat[i][:num_entities]
-        loss = hungarian_loss(pred, target, dims=2)
-        reconstruction_loss[i] = loss
+        num_entities = existence_target_flat.detach().sum(dim=-1).int()
+        reconstruction_loss = torch.zeros(entities_pred_flat.shape[0], device=entities_pred.device)
+        for i in range(entities_pred_flat.shape[0]):
+            # Extract pred and target based on actual number of entities
+            pred = entities_pred_flat[i][:num_entities[i]]
+            target = entities_target_flat[i][:num_entities[i]]
+            loss = hungarian_loss(pred, target, dims=2)
+            reconstruction_loss[i] = loss
 
-    # Reshape reconstruction loss back to batch dimensions
-    reconstruction_loss = reconstruction_loss.view(*batch_dims)
+        # Reshape reconstruction loss back to batch dimensions
+        reconstruction_loss = reconstruction_loss.view(*batch_dims)
+
+    # If the decoder has positional outputs
+    elif reconstruction_loss_type == 'positional':
+        reconstruction_loss = mse_loss(entities_pred, entities_target, dims=1)
+        reconstruction_loss = reconstruction_loss * existence_target  # Only consider existing entities
+        reconstruction_loss = reconstruction_loss.mean(dim=-1)  # Average over entities
 
     # Compute binary cross-entropy loss for existence probabilities
     existence_loss = F.binary_cross_entropy_with_logits(